+++ app/src/main/java/com/example/moviesearch/App.kt
@@ -1,44 +1,19 @@
 package com.example.moviesearch
 
 import android.app.Application
-import data.ApiConstants
-import data.MainRepository
-import data.TmdbApi
-import domain.Interactor
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
+import di.DI
+import org.koin.android.ext.koin.androidContext
+import org.koin.android.ext.koin.androidLogger
+import org.koin.core.context.startKoin
 
 class App : Application() {
 
-    private lateinit var repo: MainRepository
-    lateinit var interactor: Interactor
-    private lateinit var retrofitService:TmdbApi
-
     override fun onCreate() {
         super.onCreate()
-        instance = this
-        repo = MainRepository()
-
-        val okHttpClient = OkHttpClient.Builder().callTimeout(30, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS).addInterceptor(HttpLoggingInterceptor().apply {
-                if (BuildConfig.DEBUG) {
-                    level = HttpLoggingInterceptor.Level.BASIC
-                }
-            }).build()
-
-        val retrofit = Retrofit.Builder().baseUrl(ApiConstants.BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create()).client(okHttpClient).build()
-
-        retrofitService = retrofit.create(TmdbApi::class.java)
-
-        interactor = Interactor(retrofitService)
-    }
-
-    companion object {
-        lateinit var instance: App
-            private set
+        startKoin {
+            androidContext(this@App)
+            androidLogger()
+            modules(listOf(DI.mainModule))
+        }
     }
 }
